+++ srbuild/cli/cli.py
@@ -27,7 +27,7 @@ elif args.verbose:
 # Sets up the the command-line interface for the given project/generator combination.
 # When no profile(s) are specified, default_profile will be used.
 # TODO: Make default profile a list, and [] should correspond to all profiles.
-def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="debug"):
+def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profiles=["debug", "release"]):
     generator = GeneratorType(project)
 
     def needs_configure(func, *args, **kwargs):
@@ -40,7 +40,7 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
     # Returns the captured output
     # TODO: Subprocess management needs to be centralized somewhere.
     def _check_returncode(result: subprocess.CompletedProcess) -> str:
-        terminal_width, _ = os.get_terminal_size(0)
+        terminal_width, _ = shutil.get_terminal_size()
         output = f"\n{' Captured stdout '.center(terminal_width, '=')}\n{result.stdout.decode(sys.stdout.encoding)}\n{' Captured stderr '.center(terminal_width, '=')}\n{result.stderr.decode(sys.stdout.encoding)}"
         if result.returncode:
             G_LOGGER.critical(f"Failed with:{output}")
@@ -66,12 +66,21 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
     # Given argparse's args struct, parses out profile flags, and returns a list of profile names included.
     # Falls back to returning the default profile.
     def _select_profile_names(args) -> List[str]:
-        return [prof_name for prof_name in project.profiles.keys() if getattr(args, prof_name)] or [default_profile]
+        return [prof_name for prof_name in project.profiles.keys() if getattr(args, prof_name)] or default_profiles
 
     def _build_targets(targets: List[ProjectTarget], prof_names: List[str]):
         G_LOGGER.info(f"Building targets: {[target.name for target in targets]} for profiles: {prof_names}")
         G_LOGGER.debug(f"Targets: {targets}")
         _check_returncode(generator.build(targets, profiles=prof_names))
+
+    def help_targets(args):
+        targets = _select_targets(args)
+        for target in targets:
+            G_LOGGER.info(f"{target}. Available Profiles:")
+            for prof, node in target.items():
+                G_LOGGER.info(f"\tProfile: {prof}. Path: {node.path}.")
+                if node.install_path:
+                    G_LOGGER.info(f"\t\tInstalls to: {node.install_path}")
 
     def configure(args):
         generator.generate()
@@ -87,31 +96,48 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
         if args.target not in project.executables:
             G_LOGGER.critical(f"Could not find target: {args.target} in project executables. Note: Available targets are: {list(project.executables.keys())}")
         target = project.executables[args.target]
-        [prof_name] = _select_profile_names(args)
+        prof_name = _select_profile_names(args)[0]
         _build_targets([target], [prof_name])
-        G_LOGGER.info(f"Running target: {targets.name}, for profile: {prof_names}:\n{targets[prof_names].path}")
-        G_LOGGER.log(_check_returncode(subprocess.run([targets[prof_names].path], capture_output=True)))
+        G_LOGGER.info(f"Running target: {target}, for profile: {prof_name}:\n{target[prof_name].path}")
+        G_LOGGER.log(_check_returncode(subprocess.run([target[prof_name].path], capture_output=True)))
 
-    # TODO: FIXME: This will not work if user does not have permission to write to path
     @needs_configure
     def install(args):
-        # Select all registered installation targets by default.
-        registered_targets = [target for target in (list(project.executables.values()) + list(project.libraries.values())) if target.install_dir]
-        targets = _select_targets(args) if args.targets else registered_targets
-        [prof_name] = _select_profile_names(args)
-        _build_targets(targets, [prof_name])
+        # Select all targets by default.
+        targets = _select_targets(args)
+        prof_names = _select_profile_names(args)
+        _build_targets(targets, prof_names)
         for target in targets:
-            # Create the required directory, then install
-            if not target.install_dir:
-                G_LOGGER.critical(f"Could not find an installation entry for target: {target.name}. Have you specified one with project.install()? Note: Installation entries: {registered_targets}")
-            os.makedirs(target.install_dir, exist_ok=True)
-            path = os.path.join(target.install_dir, target[prof_name].name)
-            G_LOGGER.info(f"Installing target: {target.name}, for profile: {prof_name} to {path}")
-            shutil.copyfile(target[prof_name].path, path)
+            for prof_name in prof_names:
+                if prof_name in target:
+                    install_path = target[prof_name].install_path
+                    if install_path:
+                        os.makedirs(os.path.dirname(install_path), exist_ok=True)
+                        G_LOGGER.info(f"Installing target: {target}, for profile: {prof_name} to {install_path}")
+                        try:
+                            shutil.copyfile(target[prof_name].path, install_path)
+                        except PermissionError:
+                            G_LOGGER.critical(f"Could not write to {install_path}. Do you have sufficient privileges?")
+                    else:
+                        G_LOGGER.warning(f"No installation path is specified for target: {target} for profile: {prof_name}")
+                else:
+                    G_LOGGER.warning(f"Target: {target} does not exist for profile: {prof_name}")
 
-    # TODO: Uninstall
     def uninstall(args):
-        pass
+        targets = _select_targets(args)
+        prof_names = _select_profile_names(args)
+        if not args.force:
+            G_LOGGER.warning(f"Uninstall dry-run, will not remove files without -f/--force.")
+        for target in targets:
+            for prof_name in prof_names:
+                if prof_name in target:
+                    install_path = target[prof_name].install_path
+                    if install_path and os.path.exists(install_path):
+                        if args.force:
+                            G_LOGGER.info(f"Removing {install_path}")
+                            os.remove(install_path)
+                        else:
+                            G_LOGGER.info(f"Would remove: {install_path}")
 
     def clean(args):
         # TODO(3): Finish implementation, add per-target cleaning.
@@ -136,17 +162,23 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
 
     def _add_profile_args(parser_like, verb: str):
         for prof_name in project.profiles.keys():
-            parser_like.add_argument(f"--{prof_name}", help=f"{verb} targets using the {prof_name} profile", action="store_true")
+            parser_like.add_argument(f"--{prof_name}", help=f"{verb} targets for the {prof_name} profile", action="store_true")
 
     # By setting defaults, each subparser automatically invokes a function to execute it's actions.
     subparsers = parser.add_subparsers()
+
+    # List
+    help_targets_parser = subparsers.add_parser("targets", help="Display information about targets", description="Display information about the targets in this project")
+    help_targets_parser.add_argument("targets", nargs='*', help="Targets to display. By default, displays help information for all targets.", default=[])
+    help_targets_parser.set_defaults(func=help_targets)
+
     # Configure
     configure_parser = subparsers.add_parser("configure", help="Generate build configuration files", description="Generate a build configuration file for the project")
     configure_parser.set_defaults(func=configure)
 
     # Build
     build_parser = subparsers.add_parser("build", help="Build project targets", description="Build one or more project targets")
-    build_parser.add_argument("targets", nargs='*', help="Targets to build. By default, builds all targets for the default profile.", default=[])
+    build_parser.add_argument("targets", nargs='*', help="Targets to build. By default, builds all targets for the default profiles.", default=[])
     _add_profile_args(build_parser, "Build")
     build_parser.set_defaults(func=build)
 
@@ -159,13 +191,19 @@ def cli(project: Project, GeneratorType: type=RBuildGenerator, default_profile="
 
     # Install
     install_parser = subparsers.add_parser("install", help="Install project targets", description="Install one or more project targets")
-    install_parser.add_argument("targets", nargs='*', help="Targets to install. By default, installs all targets for the default profile.", default=[])
+    install_parser.add_argument("targets", nargs='*', help="Targets to install. By default, installs all targets for the default profiles.", default=[])
+    _add_profile_args(install_parser, "Install")
     install_parser.set_defaults(func=install)
-    install_profile_group = install_parser.add_mutually_exclusive_group()
-    _add_profile_args(install_profile_group, "Install")
+
+    # Uninstall
+    uninstall_parser = subparsers.add_parser("uninstall", help="Uninstall project targets", description="Uninstall one or more project targets")
+    uninstall_parser.add_argument("-f", "--force", help="Remove targets. Without this flag, uninstall will only do a dry-run", action="store_true")
+    uninstall_parser.add_argument("targets", nargs='*', help="Targets to uninstall. By default, uninstalls all targets for the default profiles.", default=[])
+    _add_profile_args(uninstall_parser, "Uninstall")
+    uninstall_parser.set_defaults(func=uninstall)
 
     # Clean
-    clean_parser = subparsers.add_parser("clean", help="Clean project targets", description="Clean one or more project targets. By default, cleans all targets for the default profile.")
+    clean_parser = subparsers.add_parser("clean", help="Clean project targets", description="Clean one or more project targets. By default, cleans all targets for the default profiles.")
     clean_parser.add_argument("--nuke", help="The nuclear option. Removes the entire build directory, including all targets for all profiles, meaning that the project must be reconfigured before subsequent builds.", action="store_true")
     _add_profile_args(clean_parser, "Clean")
     clean_parser.set_defaults(func=clean)