+++ srbuild/project/profile.py
@@ -11,7 +11,7 @@ def _file_suffix(path: str, suffix: str, ext: str = None) -> str:
     split = os.path.splitext(os.path.basename(path))
     basename = split[0]
     ext = ext or split[1]
-    suffixed = f"{basename}.{suffix}" + (ext if ext else "")
+    suffixed = f"{basename}{suffix}{(ext if ext else '')}"
     G_LOGGER.verbose(f"_file_suffix received path: {path}, split into {split}. Using suffix: {suffix}, generated final name: {suffixed}")
     return suffixed
 
@@ -19,10 +19,11 @@ def _file_suffix(path: str, suffix: str, ext: str = None) -> str:
 # Profiles can have default properties that are applied to each target within.
 # TODO: Add compiler/linker as a property of Profile.
 class Profile(object):
-    def __init__(self, flags: BuildFlags, build_dir: str):
+    def __init__(self, flags: BuildFlags, build_dir: str, suffix):
         self.flags = flags
         self.build_dir = build_dir
         self.graph = Graph()
+        self.suffix = suffix
 
     # libs can contain either Nodes from this graph, or paths to libraries, or names of libraries
     # This cannot be called with a target_config that has not been
@@ -37,7 +38,7 @@ class Profile(object):
             # include_dirs change, it means the file is stale, so name collisions don't matter (i.e. OK to overwrite.)
             # TODO: Maybe push signature generation into Generator.
             obj_sig = compiler.signature(source_node.path, include_dirs, flags)
-            obj_path = os.path.join(self.build_dir, _file_suffix(source_node.path, obj_sig, ".o"))
+            obj_path = os.path.join(self.build_dir, _file_suffix(source_node.path, f".{obj_sig}", ".o"))
             # User defined includes are always prepended the ones deduced for SourceNodes.
             obj_node = CompiledNode(obj_path, source_node, compiler, include_dirs, flags)
             object_nodes.append(self.graph.add(obj_node))
@@ -50,6 +51,7 @@ class Profile(object):
         # Finally, add the actual linked node
         input_paths = [node.path for node in object_nodes]
         linked_sig = linker.signature(input_paths, libs, lib_dirs, flags)
-        linked_path = os.path.join(self.build_dir, _file_suffix(basename, linked_sig))
-        linked_node = LinkedNode(linked_path, object_nodes + lib_nodes, linker, libs, lib_dirs, flags, name=basename)
+        linked_path = os.path.join(self.build_dir, _file_suffix(basename, f".{linked_sig}"))
+        display_name = _file_suffix(basename, self.suffix)
+        linked_node = LinkedNode(linked_path, object_nodes + lib_nodes, linker, libs, lib_dirs, flags, name=display_name)
         return self.graph.add(linked_node)